Change float to double for precision in calculations
Changed variable types from float to double for better precision in comparisons. Updated format specifiers in printf statements accordingly.

+++ arkusz_2/double_rownosc.c
@@ -2,7 +2,7 @@
 #include <math.h>
 
 int main() {
-float a,b,c;
+double a,b,c;  // mozna porownac na float vs double
 a=0.1;
 b=0.2;
 c=0.3;
@@ -11,12 +11,12 @@ if ((a+b)==c){
 } else{
    printf("a+b == c is false \n");
 }
-printf("a == %.30f \n", a);
-printf("b == %.30f \n", b);
-printf("a +b  == %.30f \n", a+ b);
-printf("c == %.30f \n", c);
+printf("a == %.30lf \n", a);
+printf("b == %.30lf \n", b);
+printf("a +b  == %.30lf \n", a+ b);
+printf("c == %.30lf \n", c);
 float DELTA=1e-10;
-printf("DELTA== %.30f \n", DELTA);
+printf("DELTA== %.30lf \n", DELTA);
 if ( fabs((a+b)-c)<DELTA ){
     printf("equal at specified DELTA.");
 } else {
@@ -24,3 +24,4 @@ if ( fabs((a+b)-c)<DELTA ){
 }
 return 0;
 }
+
